+++ dal/redis/video.go
@@ -7,16 +7,16 @@ import (
 )
 
 type redisVideo struct {
-	Id            int64  `json:"id" redis:"id"`
-	UserId        int64  `json:"user_id" redis:"user_id"`
+	Id            int    `json:"id" redis:"id"`
+	UserId        int    `json:"user_id" redis:"user_id"`
 	PlayUrl       string `json:"play_url" redis:"play_url"`
 	CoverUrl      string `json:"cover_url" redis:"cover_url"`
-	FavoriteCount int64  `json:"favorite_count" redis:"favorite_count"`
-	CommentCount  int64  `json:"comment_count" redis:"comment_count"`
+	FavoriteCount int    `json:"favorite_count" redis:"favorite_count"`
+	CommentCount  int    `json:"comment_count" redis:"comment_count"`
 	Title         string `json:"title" redis:"title"`
 }
 
-func VideoIsExists(ctx context.Context, vids ...int64) int64 {
+func VideoIsExists(ctx context.Context, vids ...int) int64 {
 	keys := make([]string, len(vids))
 	for i, vid := range vids {
 		keys[i] = consts.GetVideoMsgKey(vid)
@@ -25,8 +25,8 @@ func VideoIsExists(ctx context.Context, vids ...int64) int64 {
 }
 
 func SetVideoMessage(ctx context.Context, video *model.Video) (ok bool) {
-	return HSet(ctx, consts.GetVideoMsgKey(int64(video.ID)), &redisVideo{
-		Id:            int64(video.ID),
+	return HSet(ctx, consts.GetVideoMsgKey(int(video.ID)), &redisVideo{
+		Id:            int(video.ID),
 		UserId:        video.UserId,
 		PlayUrl:       video.PlayUrl,
 		CoverUrl:      video.CoverUrl,
@@ -36,15 +36,15 @@ func SetVideoMessage(ctx context.Context, video *model.Video) (ok bool) {
 	})
 }
 
-func IncrVideoField(ctx context.Context, vid int64, field string, incr int64) (ok bool) {
+func IncrVideoField(ctx context.Context, vid int, field string, incr int64) (ok bool) {
 	return HIncr(ctx, consts.GetVideoMsgKey(vid), field, incr)
 }
 
-func GetVideoFields(ctx context.Context, vid int64, field ...string) []interface{} {
+func GetVideoFields(ctx context.Context, vid int, field ...string) []interface{} {
 	return HMGet(ctx, consts.GetVideoMsgKey(vid), field...)
 }
 
-func GetVideoMessage(ctx context.Context, vid int64) (*model.Video, error) {
+func GetVideoMessage(ctx context.Context, vid int) (*model.Video, error) {
 	key := consts.GetVideoMsgKey(vid)
 	value := HGetAll(ctx, key)
 	video := new(model.Video)
